+++ src/compliance_tests/TestIso_7_5_6.cpp
@@ -78,7 +78,7 @@ class TestIso_7_5_6 : public test::TestBase
         void ConfigureTest()
         {
             FillTestVariants(VariantMatchType::CommonAndFd);
-            for (int i = 0; i < 3; i++)
+            for (size_t i = 0; i < 3; i++)
             {
                 AddElemTest(TestVariant::Common, ElemTest(i + 1, FrameKind::Can20));
                 AddElemTest(TestVariant::CanFdEna, ElemTest(i + 1, FrameKind::CanFd));
@@ -119,7 +119,7 @@ class TestIso_7_5_6 : public test::TestBase
             drv_bit_frm->InsertPasErrFrm(7, BitKind::Data);
             mon_bit_frm->InsertPasErrFrm(7, BitKind::Data);
 
-            int bit_to_corrupt;
+            size_t bit_to_corrupt;
             if (elem_test.index_ == 1)
                 bit_to_corrupt = 1;
             else if (elem_test.index_ == 2)
@@ -129,7 +129,7 @@ class TestIso_7_5_6 : public test::TestBase
 
             Bit *corrupted_bit = drv_bit_frm->GetBitOf(bit_to_corrupt,
                                     BitKind::ErrDelim);
-            int bit_index = drv_bit_frm->GetBitIndex(corrupted_bit);
+            size_t bit_index = drv_bit_frm->GetBitIndex(corrupted_bit);
             corrupted_bit->val_ = BitVal::Dominant;
 
             drv_bit_frm->InsertPasErrFrm(bit_index + 1);
